Add change payment method action
