+++ MercadoPagoSDK/MercadoPagoSDK/One tap/UI/PXOneTapViewModel.swift
@@ -36,66 +36,3 @@ final class PXOneTapViewModel: PXReviewViewModel {
         MPXTracker.sharedInstance.trackScreen(screenId: screenId, screenName: screenName, properties: properties)
     }
 }
-
-// One tap implementations.
-extension PXOneTapViewModel {
-
-    func getPaymentMethodComponent(withAction: PXComponentAction?) -> PXPaymentMethodComponent? {
-
-        guard let pm = paymentData.getPaymentMethod() else {
-            return nil
-        }
-
-        let paymentMethodName = pm.name ?? ""
-        let image = PXImageService.getIconImageFor(paymentMethod: pm)
-        var title = NSAttributedString(string: "")
-        var subtitle: NSAttributedString? = nil
-        var cftText: NSAttributedString? = nil
-        var noInterestText: NSAttributedString? = nil
-        var action = withAction
-        let backgroundColor = ThemeManager.shared.whiteColor()
-        let lightLabelColor = ThemeManager.shared.labelTintColor()
-        let boldLabelColor = ThemeManager.shared.boldLabelTintColor()
-
-        if pm.isCard {
-            if let lastFourDigits = (paymentData.token?.lastFourDigits) {
-                let text = "\(paymentMethodName) .... \(lastFourDigits)"
-                title = text.toAttributedString()
-            } else if let card = paymentOptionSelected as? CustomerPaymentMethod {
-                if let lastFourDigits = card.getCardLastForDigits() {
-                    let text: String = "\(paymentMethodName) .... \(lastFourDigits)"
-                    title = text.toAttributedString()
-                }
-            }
-        } else {
-            title = paymentMethodName.toAttributedString()
-        }
-
-        if let payerCost = paymentData.getPayerCost() {
-            // Installments.
-            let numberOfInstallments: Int = payerCost.installments
-            let installmentAmount: Double = payerCost.installmentAmount
-            let currency: Currency = MercadoPagoContext.getCurrency()
-            let amountFormatted: String = Utils.getAmountFormatted(amount: installmentAmount, thousandSeparator: currency.getThousandsSeparatorOrDefault(), decimalSeparator: currency.getDecimalSeparatorOrDefault(), addingCurrencySymbol: currency.getCurrencySymbolOrDefault(), addingParenthesis: false)
-            let subtitleDisplayText = "\(numberOfInstallments)x \(amountFormatted)"
-            subtitle = subtitleDisplayText.toAttributedString()
-
-            // Interest.
-            if !payerCost.hasInstallmentsRate() {
-                 noInterestText = "Sin interés".localized.toAttributedString()
-            }
-
-            // CFT.
-            if payerCost.hasCFTValue(), let cftValue = payerCost.getCFTValue() {
-                cftText = cftValue.toAttributedString()
-            }
-        }
-
-        if !self.reviewScreenPreference.isChangeMethodOptionEnabled() {
-            action = nil
-        }
-
-        let props = PXPaymentMethodProps(paymentMethodIcon: image, title: title, subtitle: subtitle, descriptionTitle: noInterestText, descriptionDetail: cftText, disclaimer: nil, action: action, backgroundColor: backgroundColor, lightLabelColor: lightLabelColor, boldLabelColor: boldLabelColor)
-        return PXPaymentMethodComponent(props: props)
-    }
-}
